Use name moved to appbar
