+++ src/components/app.js
@@ -8,16 +8,9 @@
 import React from 'react';
 import {useDispatch, useSelector} from 'react-redux'
 import {Redirect, Route, Switch, useHistory} from 'react-router-dom';
-import {FormattedMessage} from 'react-intl';
 
 import {createMuiTheme, makeStyles, ThemeProvider} from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import Typography from '@material-ui/core/Typography';
-import BrightnessHighIcon from '@material-ui/icons/BrightnessHigh';
-import BrightnessLowIcon from '@material-ui/icons/BrightnessLow';
-import IconButton from '@material-ui/core/IconButton';
 import Box from '@material-ui/core/Box';
 import Drawer from '@material-ui/core/Drawer';
 
@@ -26,15 +19,20 @@ import SingleLineDiagram from './single-line-diagram';
 import NetworkMap from './network/network-map';
 import StudyManager from './study-manager';
 import Network from './network/network';
-import {ReactComponent as PowsyblLogo} from '../images/powsybl_logo.svg';
+import TopBar from './top-bar';
 import {
     loadNetworkSuccess,
     loadVoltageLevelDiagramSuccess,
     openStudy,
-    removeVoltageLevelDiagram,
-    selectDarkTheme
+    removeVoltageLevelDiagram
 } from '../redux/actions'
-import {fetchLines, fetchSubstationPositions, fetchSubstations, fetchVoltageLevelDiagram, fetchLinePositions} from '../utils/rest-api'
+import {
+    fetchLinePositions,
+    fetchLines,
+    fetchSubstationPositions,
+    fetchSubstations,
+    fetchVoltageLevelDiagram
+} from '../utils/rest-api'
 
 const lightTheme = createMuiTheme({
     palette: {
@@ -53,9 +51,6 @@ const darkTheme = createMuiTheme({
 const drawerWidth = 240;
 
 const useStyles = makeStyles(theme => ({
-    appBar: {
-        zIndex: theme.zIndex.drawer + 1,
-    },
     drawer: {
         width: drawerWidth,
         flexShrink: 0,
@@ -65,18 +60,6 @@ const useStyles = makeStyles(theme => ({
     },
     grow: {
         flexGrow: 1,
-    },
-    logo: {
-        width: 48,
-        height: 48,
-    },
-    title: {
-        marginLeft: 18
-    },
-    grid: {
-        flexGrow: 1,
-        paddingLeft: theme.spacing(2),
-        bgcolor: "background.default"
     },
     toolbar: theme.mixins.toolbar,
     main: {
@@ -106,15 +89,11 @@ const App = () => {
 
     const dark = useSelector(state => state.darkTheme);
 
+    const useName = useSelector(state => state.useName);
+
     const history = useHistory();
 
     const classes = useStyles();
-
-    const useName = useSelector(state => state.useName);
-
-    function switchTheme() {
-        dispatch(selectDarkTheme(!dark));
-    }
 
     function resetStudy(studyName) {
         dispatch(removeVoltageLevelDiagram());
@@ -158,23 +137,6 @@ const App = () => {
             });
     }
 
-    function createAppBar() {
-        return (
-            <AppBar position="static" color="default" className={classes.appBar}>
-                <Toolbar>
-                    <PowsyblLogo className={classes.logo}/>
-                    <Typography variant="h6" className={classes.title}>
-                        <FormattedMessage id="appName"/>
-                    </Typography>
-                    <div className={classes.grow} />
-                    <IconButton aria-label="Change theme" color="inherit" onClick={() => switchTheme()}>
-                        { dark ? <BrightnessLowIcon /> : <BrightnessHighIcon /> }
-                    </IconButton>
-                </Toolbar>
-            </AppBar>
-        )
-    }
-
     function createNetworkPane() {
         return (
             <div>
@@ -186,7 +148,6 @@ const App = () => {
                 </Drawer>
                 <NetworkMap network={ network }
                             labelsZoomThreshold={8}
-                            useName={useName}
                             initialPosition={[2.5, 46.6]}
                             initialZoom={6}
                             onSubstationClick={ (id, name) => showVoltageLevelDiagram(id, name) } />
@@ -219,7 +180,7 @@ const App = () => {
         <ThemeProvider theme={dark ? darkTheme : lightTheme}>
             <React.Fragment>
                 <CssBaseline />
-                { createAppBar() }
+                    <TopBar />
                 { createApp() }
             </React.Fragment>
         </ThemeProvider>
