+++ src/components/network/network-explorer.js
@@ -14,17 +14,12 @@ import TextField from '@material-ui/core/TextField';
 
 import Network from "./network";
 import {useDispatch, useSelector} from "react-redux";
-import {toggleUseNameState} from "../../redux/actions";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Grid from "@material-ui/core/Grid";
-import Switch from "@material-ui/core/Switch";
 import {useIntl} from "react-intl";
 
 const itemSize = 35;
 
 const NetworkExplorer = (props) => {
     const intl = useIntl();
-    const dispatch = useDispatch();
     const useName = useSelector(state => state.useName);
     const diagram = useSelector(state => state.diagram);
 
@@ -59,31 +54,9 @@ const NetworkExplorer = (props) => {
         }));
     };
 
-    const handleToggleUseName = () => {
-        dispatch(toggleUseNameState());
-    };
-
-
     return (
         <div>
-            <Grid container spacing={2}>
-                <Grid item xs={12}>
-                    <FormControlLabel
-                        control={
-                            <Switch
-                                checked={useName}
-                                onChange={handleToggleUseName}
-                                value={useName}
-                                color="primary"
-                            />
-                        }
-                        label="Use name"
-                    />
-                </Grid>
-                <Grid item xs={12}>
-                    <TextField id="standard-basic" label={searchMsg} onChange={filter} fullWidth={true}/>
-                </Grid>
-            </Grid>
+            <TextField id="standard-basic" label={searchMsg} onChange={filter} fullWidth={true}/>
             <FixedSizeList
                 height={29 * itemSize}
                 itemCount={filteredVoltageLevels.length}
