+++ src/components/study-manager.js
@@ -183,7 +183,7 @@ const StudyManager = (props) => {
             <Grid container spacing={2} className={classes.grid}>
             {
                 studies.map(study =>
-                    <Grid item xs={3} key={study.studyName}>
+                    <Grid item xs={12} sm={6} md={3} key={study.studyName}>
                         <StudyCard study={study} onClick={() => props.onStudyClick(study.studyName)}/>
                     </Grid>
                 )
